+++ setuptools/config/legacy_setupcfg.py
